# resources/skg-if-openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0-alpha
  title: SKG-IF OpenAPI
  termsOfService: 'https://skg-if.github.io/interoperability-framework/'
  contact:
    url: 'https://github.com/skg-if'
  license:
    name: CC-BY
    url: 'http://creativecommons.org/licenses/by/4.0/'
  x-logo:
    url: 'https://www.rd-alliance.org/wp-content/uploads/2024/04/RDA_Logotype_CMYK.png'
  description: |
    OpenAPI end points specifications for [SKG-IF](https://skg-if.github.io/interoperability-framework/)

    Last version of the SKG-IF OpenAPI is available at : https://skg-if.github.io/api/

    # Note for implementers :
      - Please refer to [SKG-IF OpenAPI Implementer documentation](https://docs.google.com/document/d/1t7b7h28UTtM56Sda4NGJIp0hnQfGbcVVGn12fny9wfI/edit?usp=sharing)
      - [RDA SKG-IF WG](https://skg-if.github.io/) does not provide any official implementation of the current specification.
      - Recommended OpenAPI viewer :
        - StopLight/Elements Viewer [Current/last SKG-IF OpenAPI version](https://elements-demo.stoplight.io/?spec=https://w3id.org/skg-if/api/skg-if-openapi.yaml).
        - Swagger : [Current/last SKG-IF OpenAPI version](https://editor-next.swagger.io/)
...
servers:
  - url: '{protocol}://{server}:{port}/{skg_if_api_path}'
    variables:
      protocol:
        enum:
          - "https"
          - "http"
        default: "https"
        description: service provider protocol.
      server:
        default: example.com
...
      port:
        default: "8443"
...
paths:
...
  '/organisations/{short_local_identifier}':
    get:
      tags:
          - Organisation
      summary: Get organisation by id
      description: |
          Get `organisation` by id.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation) .\
      operationId: getOrganisationById
      parameters:
        - $ref : '#/components/parameters/shortLocalIdPathParam'
      responses:
        '200':
          description: Success
          content:
            # not 'application/json-ld' to be compatible with StopLight PRISM tool
            application/json:
              schema:
                required: ['@context','@graph']
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  "@graph":
                    type: array
                    minItems: 1
                    maxItems: 1
                    items:
                      $ref: "#/components/schemas/Organisation"
              examples:
                OrgBrownUniversity:
                  summary: "Brown University"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"
                    "@graph":
                      - local_identifier: http://example.com/skg-if/api/organisations/otf_session12324_org_1124
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/05gq02987
                        name: Brown University.
                        short_name: BU
                        country: US
                        types:
                          - education
                OrgJohnCarterBrownLibrary:
                  summary: "John Carter Brown Library - Brown University"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"
                    "@graph":
                      - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1125
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/05gq02987
                          - scheme: ror
                            value: https://ror.org/0274ane14
                        name: John Carter Brown Library, Brown University, Providence RI, US
                        short_name: BU
                        country: US
                        types:
                          - archive
                OrgLabScienceNoId:
                  summary: "Lab of science"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"
                    "@graph":
                      - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1126
                        entity_type: 'organisation'
                        name: Lab of science, unknown University, Antarctica
                OrgEuropeanCommission:
                  summary: "European Commission Funder"
                  value :
                    "@context":
                      - "https://w3id.org/skg-if/context/1.1.0/skg-if.json"
                      - "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json"
                      - "@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"
                    "@graph":
                      - local_identifier: http://example.com/skg-if/api/organisations/otf_session12324_org_1127
                        entity_type: 'organisation'
                        identifiers:
                          - scheme: ror
                            value: https://ror.org/00k4n6c32
                        name: "European Commission"
                        short_name: EU
                        types:
                          - funder
        '404':
          description: |
            Error if entity does not exists
  '/organisations':
    get:
      tags:
          - Organisation
      summary: Get list of organisation.
      operationId: getOrganisations
      description: |
          Get a list of `organisation`. See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation) .

          __Attribute filters__ \
          You can filter using these attributes of the `organisation` object
          | Filter key| Attribute filter example |
          | -------- | -------- |
          | identifiers.id | `identifiers.id:05gq02987` |
          | identifiers.scheme | `identifiers.scheme:ror` |
          | name | `name:Brown University` |
          | short_name |  `short_name:BU` |
          | website | `website:https://www.brown.edu` |
          | country | `country:US` |

          __Convenience filters__ \
          These filters aren't attributes of the `organization` object, but they're handy for solving some common use cases.

          | Filter key | Filter value | List returned | Attribute filter example |
          | -------- | -------- | ------- | ------- |
          | cf.search.name | a string | `organisation` with name containing the filter value | `cf.search.name:Brown` |

          __Query parameter syntax examples__ \

           * `/organisations?filter=identifiers.scheme:ror,identifiers.id:05gq02987`
           * `/organisations?filter=cf.search.name:Brown`
           * `/organisations?filter=cf.search.name:Brown&page=1&page_size=5`

      parameters:
        - name: filter
          in: query
          description: filter. Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
          examples :
            rorid_ex:
              value: identifiers.scheme:ror,identifiers.id:05gq02987
              summary : Get all organisation with identifier scheme 'ror' and id '05gq02987'
            name_ex:
              value: cf.search.name:Brown
              summary : Get all organisation with name containing 'Brown'
          schema:
            type: string
            pattern: '^(,?.+:.+)*$'
        - $ref : '#/components/parameters/pageQueryParam'
        - $ref : '#/components/parameters/pageSizeQueryParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  "@context":
                    type: array
                    description: JSON-LD context # not displayed in Stoplight elements
                    minItems: 3 # http://../skg-if.json, http://../skg-if-api.json and @base
                    uniqueItems: true
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/JsonLdCtxDataModel'
                        - $ref: '#/components/schemas/JsonLdCtxSearch'
                        - $ref: '#/components/schemas/JsonLdCtxBaseOrMore'
                    examples:
                      - ["https://w3id.org/skg-if/context/1.1.0/skg-if.json", "https://w3id.org/skg-if/context/1.0.0/skg-if-api.json", {"@base" : "https://w3id.org/skg-if/sandbox/my-skg-acronym/"}]
                  meta:
                    type: object
                    $ref: '#/components/schemas/Meta'
                  "@graph":
                    type: array
                    items:
                      $ref: '#/components/schemas/Organisation'
                required: [ "@context", "meta", "@graph" ]
...
components:
  securitySchemes:
    main_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://example.com/api/oauth/dialog'
          scopes:
            'read:users': read users info
            'write:users': modify or remove users
    api_key:
      type: apiKey
      in: header
      name: api_key
    basic_auth:
      type: http
      scheme: basic
  parameters:
    shortLocalIdPathParam:
      name: short_local_identifier
      in: path
      description: |
        entity id. URL suffix used in OpenAPI get by id operations. \
        `https://[your-server-root-skg-if-url]/[products|persons|organisations|datasources|venues|topics]/{short_local_identifier}` \
        example `http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9`
      required: true
      schema:
        type: string
#    filterQueryParam: # filterQueryParam duplicated in each search endpoint ( only way to add distinct examples for each endpoint )
#      name: filter
#      in: query
#      description: |
#            Search filter. Format : Coma separated filter_name:filter_value elements ( filter_name_1:filter_value_1,filter_name_2:filter_value_2,filter_name_3:filter_value_3... ). Server side operator used is _AND_.
#      schema:
#        type: string
#        pattern: '^(,?.+:.+)*$'
    pageQueryParam:
      name: page
      in: query
      description: |
            Search page reference (page number or search iteration token).\
            Note: As a client application you should build your iterating logic using the `meta.next_page.local_identifier` and `meta.prev_page.local_identifier` URLs
      schema:
        type: string
    pageSizeQueryParam:
      name: page_size
      in: query
      description: |
            Search page size. limit search result size.
      schema:
        type: integer
  schemas:
...
    Organisation:
      type: object
      title: "Organisation"
      description:  "`organisation` object.  See definition in SKG-IF [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html) ( entity_type:organisation )."
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          required: [
            "local_identifier","entity_type"
          ]
          properties :
            entity_type:
              enum:
                - "organisation"
              type: string
              x-faker:
                fake: ["organisation"]
            identifiers:
              description: Objects representing external identifiers for the entity. See [Agent](https://skg-if.github.io/interoperability-framework/docs/agent.html#identifiers) documentation for supported schemes
              type: array
              items:
                required: [ "scheme", "value" ]
                properties:
                  scheme:
                    description: The scheme for the external identifier.
                    type: string
                    x-faker:
                      fake: ["ror"]
                    examples:
                      - 'ror'
                  value:
                    type: string
                    description: "The external identifier."
                    x-faker:
                      fake : ['{{random.alphaNumeric(10)}}']
                    examples:
                      - "https://ror.org/05gq02987"
            name:
              type: string
              x-faker: company.companyName
              examples:
                - "Brown University"
            other_names:
              description:  "Other names, perhaps in different languages, identifying an `organisation`."
              type: array
              items:
                type: string
              examples :
                - ["Brownie"]
            short_name:
              type: string
              #x-faker: company.bsAdjective
              x-faker:
                random.alpha: [{count: 4, upcase: true}]
              examples:
                - "BU"
            website:
              type: string
              description: The website URL for an Organisation
              examples:
                - "https://www.brown.edu/"
            country:
              type: string
              description: The country code expressed as [ISO 3166-1 alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2).
              pattern: '^[A-Z]{2}$'
              examples:
                - "US"
            types:
              description:  "The types of an `organisation`."
              type: array
              items:
                type: string
                enum:
                  - archive
                  - company
                  - education
                  - facility
                  - government
                  - healthcare
                  - nonprofit
                  - funder
                  - research
                  - unspecified
      examples:
        - local_identifier: 'otf_session12324_org_1124'
          entity_type: 'organisation'
          identifiers:
            - "scheme": ror
              "value": https://ror.org/05gq02987
          name: "Brown University"
          short_name: BU
          country: US
          types:
            - education
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1125
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/05gq02987
            - scheme: ror
              value: https://ror.org/0274ane14
          name: John Carter Brown Library, Brown University, Providence RI, US
          short_name: BU
          country: US
          types:
            - archive
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12325_org_1126
          entity_type: 'organisation'
          name: Lab of science, unknown University, Antarctica
        - local_identifier: http://example.com/skg-if/api/organisations/otf_session12324_org_1127
          entity_type: 'organisation'
          identifiers:
            - scheme: ror
              value: https://ror.org/00k4n6c32
          name: "European Commission"
          short_name: EU
          types:
            - funder
...
    Entity:
      type: object
      description: "Entity common fields used for entities : `product`,`person`,`organisation`,`grant`,`venue`,`datasource`"
      properties:
        local_identifier:
          type: string
          description: |
            Unique local identifier as URL. \
              `https://[your-server-root-skg-if-url]/[products|persons|organisations|datasources|venues|topics]/{short_local_identifier}` \
              example `http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9` \
            \
            If the entity does not have any stable identifier an on-the-fly identifier is returned : `otf___<session identifier>___<identifier string>`.
            -  `otf` stands for on-the-fly to explicitly clarify the local identifiers it has been created for the purpose of creating this SKG-IF document;
            -  `<session identifier>` is a string portion that enables the source to uniquely identifier the session in which such SKG-IF document has been created - e.g. it could be the current time of the software run to create the SKG expressed in milliseconds;
            -  `<identifier string>` is a string portion that identifies the particular entity in consideration.
            For instance, a possible example of such a local identifier would be : `otf___1730027051396___person-1`. An on-the-fly may not resolve with URL `/<entity_type_get_operation>/<otf_...>`.
          x-faker:
            fake : ['{{random.uuid}}']
          examples:
            - http://example.com/skg-if/api/products/c66c6-38be-4d5f-85db-d44c9
            - http://example.com/skg-if/api/products/otf___1730027051396___prod-1
            - http://example.com/skg-if/api/persons/c7777-38be-4d5f-85db-d7777
            - http://example.com/skg-if/api/persons/otf___1730027051396___person-1
        identifiers:
          type: array
          items:
            required: [ "scheme", "value" ]
            properties:
              scheme:
                type: string
                # alias list in https://skg-if.github.io/context/ver/current/skg-if.json
                # enum: ["doi","pmid","pmcid","arxiv","url","omid","crossref","openalex","orcid","viaf","ror": ,"issn": , "eissn":, "lissn","isbn","opendoar","re3data.org","fairsharing","handle"]
              value:
                type: string
        entity_type:
          type: string
    Meta:
      type: object
      required: [ "local_identifier", "entity_type" ]
      properties:
        local_identifier:
          type: string
          description: search result page id
          examples:
            - https://example.com/skg-if/api/products?filter=xxx&page=y
            - https://example.com/skg-if/api/persons?filter=xxx&page=y
        entity_type:
          type: string
          description: search result page type
          enum:
            - search_result_page
        prev_page:
          description: Previous page reference
          type: object
          $ref: '#/components/schemas/SearchResultPage'
        next_page:
          description: Next page reference
          type: object
          $ref: '#/components/schemas/SearchResultPage'
        part_of:
          description: Reference for the search result as a whole (all pages)
          type: object
          required: [ "local_identifier", "entity_type" ]
          properties:
            local_identifier:
              type: string
              description: search result id as URL
              examples:
                - https://example.com/skg-if/api/products?filter=xxx
                - https://example.com/skg-if/api/persons?filter=xxx
            entity_type:
              type: string
              description: search type. Value must be `search_result`
              enum:
                - search_result
            total_items:
              description: total number of item entities for the search
              type: integer
              x-faker:
                datatype.number: [{min: 10, max: 100}]
            first_page:
              description: First page reference of the search
              type: object
              $ref: '#/components/schemas/SearchResultPage'
            last_page:
              description: Last page reference of the search
              type: object
              $ref: '#/components/schemas/SearchResultPage'
    SearchResultPage:
        type: object
        properties:
          local_identifier:
                type: string
                description: search result page id as URL
                examples:
                  - https://example.com/skg-if/api/products?filter=xxx&page=y
                  - https://example.com/skg-if/api/persons?filter=xxx&page=y
          entity_type:
            type: string
            description: search result page type. Value must be `search_result_page`
            enum:
              - search_result_page
...
    JsonLdCtxDataModel:
      type: string
      description: "URL to SKG-IF data model context supported by the current OpenAPI `https://w3id.org/skg-if/context/1.1.0/skg-if.json`"
      enum:
        - https://w3id.org/skg-if/context/1.1.0/skg-if.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.1.0/skg-if.json']
    JsonLdCtxSearch:
      type: string
      description: "URL to SKG-IF API context (meta section) supported by the current OpenAPI `https://w3id.org/skg-if/context/1.0.0/skg-if-api.json`"
      enum:
        - https://w3id.org/skg-if/context/1.0.0/skg-if-api.json
      x-faker:
        fake: ['https://w3id.org/skg-if/context/1.0.0/skg-if-api.json']
    JsonLdCtxBaseOrMore:
      description : 'Additional context mappings ex: `{ "@base":"https://w3id.org/skg-if/sandbox/my-skg-acronym" }`'
      properties:
        "@base":
          type: string
          description : |
            Default base URL specific to the implementer.\
            common format : `https://w3id.org/skg-if/sandbox/{my-skg-acronym}` \
            example : `https://w3id.org/skg-if/sandbox/openaire/`
          x-faker:
            fake: ['https://w3id.org/skg-if/sandbox/{{random.alpha(5)}}/']
          examples:
            - "https://w3id.org/skg-if/sandbox/openaire/"
            - "https://w3id.org/skg-if/sandbox/cessda/"
            - "https://w3id.org/skg-if/sandbox/foobar/"
      additionalProperties:
        type: string
        description: "context URL of extension entities"
        x-faker:
          fake : ['http://www.example.com/skg-if/{{random.alpha(3)}}']